Fix (LP: #1765565): Don't crash on corrupted panel configs.
1. Fix the problem directly:

Catch the exception if the .desktop file is missing and return false.
This will remove launchers with missing .desktop (not just the binary).

2. Fix the problem more generally:

Ignore orphan plugin configurations. This is a plugin configuration
section which is not refered to by any panel and therefore is not used.
Some old versions of xfce4-panel seem to create these, but I can't
reproduce with the latest version.

These orphans are completely ignored by the panel. Xfpanel-switch also
ignores them, unless they are launchers with missing .desktop files.
The second part of this patch cleans them out anyway.