Save and restore launchers as is
There doesn't seem to be any good reason to exclude launchers that users
want to keep on their panel, whether these launchers are empty, or the
app they point to isn't installed at the time of saving, or for any
other reason.

What's more, this deletion was only partially done, leaving the launcher
id in '/panels/panel-N/plugin-ids', so it caused a panel error dialog to
be displayed when restoring.

This also solves the problem of link launchers that were excluded, as
they had no Exec key.

Fixes: #52