Revert "Replace deprecated exo with libxfce4ui 4.21.0"
This reverts commit 1f5d16efd7e6214ad2cd1e34a7ac22b83bc0aa84.

This won't work in general, as packages are built in containers that
contain only the dependencies needed for the build, which excludes
libxfce4ui and therefore the presence of xfce-open in this case.

Checking for the presence of xfce-open at runtime is not obvious here,
as it would require you to put shell code in the 'Exec=' field of
'.desktop' files, not to mention '.rc' files where this is not possible
at all. The only solution therefore seems to be to wait until Xfce 4.22
is released and simply replace exo-open with xfce-open at that point.