Remove wrong sanity check
The documentation for the GVariant format indicates that this format
generally corresponds to the output of Python's str() function, but that
there are additions and exceptions. Therefore, this may sometimes fail,
and it seems unnecessary to replace it with the composition of
g_variant_parse() and g_variant_print(), which are by definition
reciprocal.

Fixes: #58